refactor: 웹 어플리케이션을 위한 패키지 분리
